todo_server/pyproject.toml
[tool.poetry]
name = "todo"
version = "0.1.0"
description = ""
authors = ["ofek4454 <ofek4454@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.95.0"
tortoise-orm = {extras = ["asyncpg"], version = "^0.19.3"}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
